consistency tweaks
git-svn-id: https://svn.r-project.org/R/trunk@72229 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/include/Rinternals.h
@@ -1451,7 +1451,7 @@ SEXP R_FixupRHS(SEXP x, SEXP y);
 	void NORET R_SignalCStackOverflow(intptr_t);				\
 	int dummy;							\
 	intptr_t usage = R_CStackDir * (R_CStackStart - (uintptr_t)&dummy); \
-	if(R_CStackLimit != -1 && usage > ((intptr_t) R_CStackLimit))	\
+	if(R_CStackLimit != (uintptr_t)(-1) && usage > ((intptr_t) R_CStackLimit)) \
 	    R_SignalCStackOverflow(usage);				\
     } while (FALSE)
 #endif
