+++ src/unix/system.c
@@ -234,7 +234,7 @@ int Rf_initialize_R(int ac, char **av)
 	R_CStackStart = (uintptr_t) &i + (6000 * R_CStackDir);
     }
 #endif
-    if(R_CStackStart == -1) R_CStackLimit = -1; /* never set */
+    if(R_CStackStart == (uintptr_t)(-1)) R_CStackLimit = (uintptr_t)(-1); /* never set */
 
     /* printf("stack limit %ld, start %lx dir %d \n", R_CStackLimit,
 	      R_CStackStart, R_CStackDir); */
